requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.19
albumentations==1.4.20
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astutils==0.0.6
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
batchgenerators==0.25.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
dd==0.5.7
debugpy==1.8.7
decorator==4.4.2
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
efficientnet_pytorch==0.7.1
et_xmlfile==2.0.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.14.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h5py==3.12.1
hf-xet==1.2.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
iopath==0.1.10
ipycanvas==0.14.2
ipyevents==2.0.4
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==0.10.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.16
kagglehub==0.3.10
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.8
linecache2==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmcv==2.2.0
mmengine==0.10.5
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.1.0
multimethod==1.12
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
nibabel==5.3.2
notebook_shim==0.2.4
numpy==2.1.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
olefile==0.47
omegaconf==2.3.0
openai==1.93.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.0
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
ply==3.10
portalocker==2.10.1
pretrainedmodels==0.7.4
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==4.0.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytorch-lightning==2.4.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2025.10.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.20.0
safetensors==0.4.5
# Editable install with no version control (SAM-2==1.0)
-e /home/zheyaogao/Experiments/ESD_seg/sam2
scikit-image==0.25.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
Send2Trash==1.8.3
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.10.5
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.9.7
tinycss2==1.3.0
tokenizers==0.22.1
tomli==2.0.2
torch==2.4.1
torch-explain==1.5.1
torchaudio==2.4.1
torchmetrics==1.5.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.1
traceback2==1.4.0
traitlets==5.14.3
transformers==4.57.1
triton==3.0.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
unittest2==1.1.0
uri-template==1.3.0
urllib3==1.26.20
wandb==0.18.7
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.15
yapf==0.43.0
yarl==1.17.1

